refactor: make MSZXFile inherit from MSZFile in Python bindings
